+++ control-plane/connect-inject/annotations.go
@@ -78,6 +78,11 @@ const (
 	// Deprecated: This annotation is no longer supported.
 	annotationSyncPeriod = "consul.hashicorp.com/connect-sync-period"
 
+	// annotationUseProxyHealthCheck creates a readiness listener on the sidecar proxy and
+	// queries this instead of the application health check for the status of the application.
+	// Enable this only if the application does not support health checks.
+	annotationUseProxyHealthCheck = "consul.hashicorp.com/use-proxy-health-check"
+
 	// annotations for sidecar proxy resource limits.
 	annotationSidecarProxyCPULimit      = "consul.hashicorp.com/sidecar-proxy-cpu-limit"
 	annotationSidecarProxyCPURequest    = "consul.hashicorp.com/sidecar-proxy-cpu-request"